+++ pp/sidebarsPp.js
@@ -23,6 +23,10 @@ module.exports = {
               type: 'doc',
               id: 'integrations/plugin-packs/getting-started/how-to-guides/centreon-nsclient-tutorial'
             },
+			{
+              type: 'doc',
+              id: 'integrations/plugin-packs/getting-started/how-to-guides/telegraf'
+            },
             {
               type: 'doc',
               id: 'integrations/plugin-packs/getting-started/how-to-guides/collect-openmetrics'
@@ -2014,8 +2018,9 @@ module.exports = {
         },
         {
           type: 'doc',
-          id: 'integrations/plugin-packs/procedures/network-switchs-symbol-wing-snmp'
+          id: 'integrations/plugin-packs/procedures/network-symbol-wing-snmp'
         },
+
         {
           type: 'doc',
           id: 'integrations/plugin-packs/procedures/network-teldat-snmp'